+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## Unreleased
 
+### Fixed
+
+* Change worker OOM handling strategy
+
 ## 1.9.4 - 2025-03-24
 
 ### Added